Implemented method to get the height of a binary tree.
Because we do not have a balanced binary tree, we cannot access the height using
known equations. In this implementation, the method calculates the height of the tree every time
it is called. For future implementation, it would be worthwhile to find a
way to deliver the height in constant time for most of the function calls.

The test cases address an empty input, single input, somewhat balanced tree, and an
unbalanced tree.