Added theoreticalHeight method to calculate minimum possible height for given tree
The theoreticalHeight method utilizes the fact that each level can store 2^i number of
nodes, where i represents the level of the tree. Here we are calculating the level
with a given number of nodes, which is log2(numItems).

OTHER CHANGES:

The test case for testing Insert method with the constructor was technically testing the
numItems value. This was renamed to TestNumItems.