Templating the UniqueClass Testcases
There was an issue with the NUnit Testing in Unique.Test regarding the TargetFramework.
The Unique target was net6.0-windows whereas the Unique.Test target was net6.0. This
caused an error to throw when trying to build the test project. Solution was to restore
project with target for net6.0-windows.

First test case written for Unique Class Constructor. The constructor will initialize a
List<uint> type with random numbers based on input parameters `size` and `max`.