Wrote test cases for DinstinctIntegers
Template for UniqueStatistics class written
Test cases for UniqueStatistics class written

Since the methods in the homework take in the same inputs and output the same
type with same expected output, I wrote a test case that takes in a list, expected output,
and a function type as parameters and compares the expected output with  the function call.
This avoids us writing redundant test cases that test the same inputs but with a different method.