Allowed StyleCop to generate warnings for XML comments in the test project and fixed bugs in test project
For StyleCop to generate warnings regarding XML formatting, a tag had to be inserted into the .csproj file.
This change was made to allow warning messages to appear and to remove Warning SA0001.

I misinterpretted the goal of the DistinctIntegers functions. My assumption was to only accept an integer as
distinct if it existed once and only once in the array. The implementation actually accepts all integers with a
unique binary equivalant, meaning it will exclude any future integer with an already existing binary equivalent.
This change was adjusted in the test cases.