Completed UI Design
I moved the UI Implementation to the namespace `Unique.Forms`. The only reasoning was
to further understand how namespaces work in C#.

The Microsoft Documentation did not separate the `UniqueForm` into a partial class like
the dotnet template did for this project. However, within UniqueForm.InitializeComponent(),
it was necessary to append a load event handler in order to call the UniqueForm_Load method.