Initialized Project
Initialized the structure for homework 3. Homework 3 will require a more in depth
gui, which the forms are located in the `Forms` file within the `Fibonacci` project.

A testing template is set up as well in `Fibonacci.Test` which utilizes NUnit.

StylCop is also included in both projects and is functioning as intended.