Restructuring Layout: Forms
Relocated the files related to Forms to its own separate file within Notepad project.
As I add more files to the project, it makes sense to have the UI elements in one file
and other unrelated files in other locations.

The structure is currently aiming to match the following tree:

Notepad/
    Forms/
    TextReader/
    .csproj
Notepad.Test/
    Test*.cs
    .csproj