Implement FibonacciTextReader Class
Implented the FibonacciTextReader Class to the specificiations of the
FibonacciTextReaderTests. The Fibonacci finds the 100th value very quickly.
Although at the cost of memory, storing the results of previous calculated values
really help speed up the algorithm. I decided to implement a ConcurrentDictionary
as it was a static member. I was thinking about writing for the future, and
if there is a possibility of threads in the future, this would protect read/write access
between instances of the class and sharing the buffer.