Fixed Access permissions to Cell object in Spreadsheet
The requirements for the project say to not allow outside sources to be able to set or see the value of the
implemented SpreadSheet Cell being used in Spreadsheet class. I felt like it is needed to make the implemented
SpreadsheetCell public as something besides Spreadsheet could eventually use it in the future. However, when the
Spreadsheet class initializes SpreadsheetCell objects, it does not allow any other class outside of it
see or modify them.