Refactored the Factory Method for Getting Nodes
In the previous homework, I simply hardcoded the list of supported operators into a dictionary.
This version does not require we add the operators we created to the dictionary. As long as there
is an `Operation` variable in the class and it inherits from `OperatorNode` it will be added to the
Dictionary when the program runs.

Addtionally, since I am using reflections now, I wrote test cases to ensure the NodeFactory is still
returning the correct Nodes and that all expected OperatorNodes are included in the Dictionary.