Added Precedence Properties to Implemented OperatorNodes
The implemented OperatorNodes will now need to have an additional static property implemented: `Precedence`.
This property will help the tree builder determine which operations have higher or lower priority.
It would be ideal to have the static properties `Precedence` and `Operation` be enforced through
some sort of Interface, but I dont think this is possible for static members.