Refactored OperatorNode and NodeFactory
Accessing the static members for the implemented `OperatorNode` classes made it so
we could not see them when using the parent class. Although we lose the ability to view the
`Operation` and `Precedence` without instantiating a class instance, this way we can enforce they be
written in any new child class of `OperatorNode`.

The `CreateNode` from `NodeFactory` was also able to be reduced due to this change, as now we
now can ensure the two properties exist in every child class of `OperatorNode`.