Implement ExpressionTree with precedence
The expression tree now supports precedence and parantheses.

For the future, it might be better to separate the factory method to support only OperatorNodes
and make another for the Variable and Constant Node classes. Although this way works and is
compact, we have to cast the `Node` to be an `OperatorNode` when it comes to finding right and left
nodes. Either way, we need to check if the node returned is null, so making a cast after is not a
big deal.