Added items to the Cell objects
In the Spreadsheet Cell, added was the ExpressionTree. It is a nullable field as some may or may not need it.
I may remove this attribute in the future as it may not be nessecary to include it in the Cell class (ie. build it
as needed). Although, we do need to unsubscribe from events, so for now it will stay.

In the Cell, added a new method that helps subscribe to events.