Added method to create a new tree.
The original design for the ExpressionTree was that the user had to create a new object with a
new expression in it in order to calculate a new expression. This reduced the level of encapsulation,
and as a result I implemented a new method, `ExpressionTree.CreateNewTree()` which creates a new tree
with a given expression.